requirements.txt
streamlit
pymoo
networkx
pandas
plotly
scikit-learn
requests
javalang
matplotlib
scipy
gitpython
numba